pyproject.toml
[project]
name = "gui-ai-ocr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.42.39",
    "fastapi>=0.128.2",
    "openai>=2.17.0",
    "openpyxl>=3.1.5",
    "opentelemetry-exporter-otlp-proto-grpc>=1.39.1",
    "opentelemetry-sdk>=1.39.1",
    "pdf2image>=1.17.0",
    "pillow>=12.1.0",
    "pydantic>=2.7,<3",
    "pymupdf>=1.26.7",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "uvicorn[standard]>=0.40.0",
]
